Move system unit tests to a separate file
Follow the convention used for other unit tests.
